user related changes
